SchermateGUI division into views
