+++ AssociateView.py
@@ -1,0 +1,58 @@
+from tkinter import *
+import UtilityView as uv
+
+# caratteristiche font
+font_size_piccolo = 20
+font_size_medio   = 33
+font_size_grande  = 80
+font_stile = "Helvetica"
+font_piccolo = (font_stile, font_size_piccolo)
+font_medio = (font_stile, font_size_medio)
+font_grande = (font_stile, font_size_grande)
+
+button_background_color = "#404040"
+active_background_color="#C8D7DC"
+root_background_color = "#708090"
+font_color = "#E0E0E0"
+
+
+class AssociateView:
+
+    # Schermata che appare dopo aver cliccato su ASSOCIA nel MENU PRINCIPALE
+    def schermata_pulsanti(closingroot, number_of_button):
+
+        closingroot.quit()
+        root = Tk()
+        root.attributes('-fullscreen', True)
+
+        frame = Frame(root)
+
+        text = Text(frame, wrap="none", bg=root_background_color)
+        vsb = Scrollbar(frame, orient="vertical", command=text.yview,width=40)
+        vsb.config(width=90)
+        text.configure(yscrollcommand=vsb.set,width=3,bg=root_background_color)
+        vsb.pack(side="left", fill="y")
+        text.pack(side ="left",fill="both",expand=True)
+
+        pulstante_uscita = Button(frame,
+                                  text="Torna al\nmenu principale",
+                                  command=lambda: root.destroy(),
+                                  bd=20,
+                                  bg=button_background_color,
+                                  font=font_piccolo,
+                                  fg=font_color,
+                                  activebackground=active_background_color)
+        pulstante_uscita.config(height=50, width=18)
+
+        #  ciclo che crea "number_of_button" pulsanti
+        for i in range(number_of_button):
+            pulsante = uv.bottom_with_text(frame, "Pulsante " + str(i + 1))
+            pulsante.pack(side=TOP, fill=BOTH)
+            text.window_create("end", window=pulsante)
+            text.insert("end", "\n")
+
+        text.configure(state="disabled")
+        frame.pack(fill="both", expand=True)
+        pulstante_uscita.pack(side=RIGHT, fill=BOTH)
+        root.mainloop()
+
