+++ BlindoMainProgramm.py
@@ -1,4 +1,3 @@
-
 from SchermateGUI import *
 
 SchermateGUI.menu_principale()
