+++ GUIkeyboard.py
@@ -23,25 +23,20 @@ result=""
 def select(value):
 
 	if value == "CANCELLA":
-		# allText = entry.get()[:-1]
-		# entry.delete(0, tkinter,END)
-		# entry.insert(0,allText)
-
-		entry.delete(len(entry.get())-1,tkinter.END)
+		entry.delete(len(entry.get())-1, tkinter.END)
 
 	elif value == "SPAZIO":
 		entry.insert(tkinter.END, ' ')
 	elif value == " Tab ":
 		entry.insert(tkinter.END, '    ')
-	#elif value == "INVIO":
-	else :
+	else:
 		entry.insert(tkinter.END,value)
+
 def close(root):
-        
-        global result
-        result = entry.get()
-        root.quit()
-        root.destroy()
+    global result
+    result = entry.get()
+    root.quit()
+    root.destroy()
 
 def HosoPop(kb):
 	key_background = "orange2"				#"#3c4987"
@@ -51,8 +46,8 @@ def HosoPop(kb):
 	key_relief = 'raised'
 	key_font = ("Helvetica", 13)
 
-	varRow = 2
-	varColumn = 0
+	var_row = 2
+	var_column = 0
 
 	for button in buttons:
 
@@ -69,7 +64,7 @@ def HosoPop(kb):
 						font=key_font,
 						padx=1, pady=1,
 						bd=5,
-						command=command).grid(row=6, column=0, columnspan=3)
+						command=command).grid(row=6, column=0, columnspan=5)
 
 		elif button == "SPAZIO":
 			tkinter.Button(
@@ -81,7 +76,7 @@ def HosoPop(kb):
 						relief=key_relief,
 						padx=1, pady=1,
 						bd=5,
-						command=command).grid(row=6, column=1, columnspan=9)
+						command=command).grid(row=6, column=2, columnspan=8)
 		elif button == "SALVA" :
 			tkinter.Button(
 						kb, text=button, width=15, heigh=6,
@@ -93,7 +88,7 @@ def HosoPop(kb):
 						relief=key_relief,
 						padx=1, pady=1,
 						bd=5,
-						command=lambda: close(kb)).grid(row=6, column=8, columnspan=3)
+						command=lambda: close(kb)).grid(row=6, column=7, columnspan=5)
 		else:
 			tkinter.Button(
 						kb, text=button, width=5, heigh=2,
@@ -105,20 +100,20 @@ def HosoPop(kb):
 						relief=key_relief,
 						padx=1, pady=1,
 						bd=6,
-						command=command).grid(row=varRow, column=varColumn)
+						command=command).grid(row=var_row, column=var_column)
 
 
-		varColumn +=1
+		var_column +=1
 
-		if varColumn > 13 and varRow == 2:
-			varColumn = 0
-			varRow+=1
-		if varColumn > 13 and varRow == 3:
-			varColumn = 0
-			varRow+=1
-		if varColumn > 13 and varRow == 4:
-			varColumn = 0
-			varRow+=1
+		if var_column > 13 and var_row == 2:
+			var_column = 0
+			var_row+=1
+		if var_column > 13 and var_row == 3:
+			var_column = 0
+			var_row+=1
+		if var_column > 13 and var_row == 4:
+			var_column = 0
+			var_row+=1
 
 def keyBoard():
 	label_font = ("Helvetica",20)
