+++ Pause.py
@@ -9,13 +9,10 @@ class Pause:
         if self.unpaused:
             PM.music.pause()
             # print("in pause")
-            self.unpaused=not self.unpaused
-            
-        else:
-            PM.music.unpause()
             # print("in play")
             self.unpaused=not self.unpaused
 
     # this method will reset the paused value to the default one, False
     def reset(self):
         self.unpaused = True
+
