+++ path_per_pc.txt
@@ -1,0 +1,18 @@
+'''
+                                        YARI'S PATH
+path_punto_accesso_chiavette = r"C:\Users\yari7\Downloads\UNIBS\IEEE\Projects\Blindo\fileAudiofromChiavetta"
+path_che_simula_la_memoria_interna_del_raspberry = r"C:\Users\yari7\Downloads\UNIBS\IEEE\Projects\Blindo\fileAudioRSPmemory"
+                                        
+                                        SAMU'S PATH
+path_punto_accesso_chiavette = r"D:\Programmi\PyCharm Projects\Chiavette"
+path_che_simula_la_memoria_interna_del_raspberry = r"D:\Programmi\PyCharm Projects\Memoria raspberry"
+
+                                        MASSI'S PATH
+path_punto_accesso_chiavette = r""
+path_che_simula_la_memoria_interna_del_raspberry = r""
+
+                                        DIEGO'S PATH
+path_punto_accesso_chiavette = r"C:\Users\Diego Berardi\Desktop\file audio blindo\punto di accesso chiavette"
+path_che_simula_la_memoria_interna_del_raspberry = r"C:\Users\Diego Berardi\Desktop\file audio blindo\simula memoria interna"
+
+'''
